refactor(backlog): move unit tests from mod.rs to tests.rs
